build: add additional exports from package.json (#37)
preinstalled snaps require additional information. If we want to be ESM
compatible this needs to be exposed on the package.json

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890

Try to provide a description, images, or anything that helps make it
clear the intent of the PR.
Thanks!
-->