fix: poc inliner
Using this snap as a preinstalled snap in the MetaMask Extension currently requires that the MetaMask Extension stringifies the bundle and icon (the manifest can actually ve imported normally). Since this snap is now intended to be preinstalled, I think this step should be the responsibility of the snap itself, not the Metamask Extension's build process.

The build system currently uses `brfs`, but I was hoping to remove it as it is slow (necessarily so, as it must statically analyze every single javascript file looking for calls to `fs.read*`) and I'm on a mission to speed up the build process.