feat: add `srpId` management (#116)
This PR is related to the SIP-30 implementation, and how we are
leveraging it for multi-SRP and Solana account syncing.

It adds one new method, `getAllPublicKeys`, that returns an array of
tuples `[EntropySourceId, SrpId][]`.
It also adds an optional param to `getPublicKey`, `entropySourceId`,
that permits querying the public key for a specific entropy source id.

With this PR, the result of `getPublicKey` ultimately becomes an `srpId`

Fixes: https://consensyssoftware.atlassian.net/browse/IDENTITY-41