feat: generate unique entropy and keys for different origins (#121)
fixes
[IDENTITY-2](https://consensyssoftware.atlassian.net/jira/software/projects/IDENTITY/boards/727?selectedIssue=IDENTITY-2)

This is a BREAKING CHANGE.

The message-signing-snap remains accessible by other origins, but it
creates unique keys for every origin.
The entropy used as a private key (and implicitly public keys /
signatures / storage_keys / etc ) will be origin specific after this
fix.
Backwards compatibility is maintained for the origins we were already
using as `initialConnections`; they are treated the same as the empty
origin, which represents calls coming from the extension or mobile app.