chore: update dependencies (#131)
Currently dependencies are out of date which cause security problems.
Since this snap is also used with gator snap the security problems get
propagated through the dependency tree.

This PR updates all the dependency to the latest version (except eslint
and zod dependencies) and fixates versions in `package.json`.

Eslint dependencies should be addressed in a separate PR since they
result in changes in the whole code base - style wise.